fix: allow foundation objective across stages
